Unify operator deployment config
realized we already have a top level field for it, and it was in common

[static]
