+++ cluster/pulumi/operator/src/flux/flux-alerts.ts
@@ -3,8 +3,8 @@
 import * as k8s from '@pulumi/kubernetes';
 import { CLUSTER_BASENAME, clusterProdLike, config } from '@canton-network/splice-pulumi-common';
 
+import { fluxConfig } from '../config';
 import { namespace } from '../namespace';
-import { operatorFluxConfig } from './config';
 import { flux } from './flux';
 
 if (clusterProdLike) {
@@ -30,7 +30,7 @@ if (clusterProdLike) {
       },
       spec: {
         type: 'slack',
-        channel: operatorFluxConfig.flux.alertSlackChannel,
+        channel: fluxConfig.alertSlackChannel,
         address: 'https://slack.com/api/chat.postMessage',
         secretRef: { name: slackToken.metadata.name },
       },
