+++ cluster/pulumi/operator/src/index.ts
@@ -4,7 +4,7 @@ import { CLUSTER_BASENAME } from '@canton-network/splice-pulumi-common';
 import { gitRepoForRef } from '@canton-network/splice-pulumi-common/src/operator/flux-source';
 import { createEnvRefs } from '@canton-network/splice-pulumi-common/src/operator/stack';
 
-import { operatorDeploymentConfig } from '../../common/src/operator/config';
+import { operatorDeploymentConfig } from './config';
 import { flux } from './flux';
 import { namespace } from './namespace';
 import { installDeploymentStack } from './stacks/deployment';
